Today's work
Wrote all the necessary code to use the data batch corrected on the WTs by subbatch - main file can now use either that or the outlier_removed (traditional one) as input. Based on this, started modifying file for correlation to Pilpel et al., 2014 and started writing my own version of Alexis' plots. Also wrote a separate script to do all the processing I need on the sample_layout from Boris at once, and write to a directory of mine in the S drive